+++ src/Classifiers/LogisticRegression.php
@@ -327,7 +327,7 @@ class LogisticRegression implements Estimator, Learner, Online, Probabilistic, R
             $this->steps[] = $loss;
 
             if ($this->logger) {
-                $this->logger->info("Epoch $epoch {$this->costFn}=$loss");
+                $this->logger->info("Epoch $epoch - {$this->costFn}: $loss");
             }
 
             if ($loss < $bestLoss) {